+++ RELEASENOTES.md
@@ -12,8 +12,9 @@ Includes minor changes that affect the results to a limited degree, specifically
 1. Removed paper related scripts/documents. These have been moved to [this GitHub repository](https://github.com/bramdr/papers). Currently only VIC related scripts/documents are included in this repository.
 2. Updated the folder structure to remove redundand data. All data is located in the folder 'Data', all scripts are located in the folder 'Scripts'.
 3. Included domain generation: single and/or multiple cell generation; single and/or multiple basin generation.
-4. Added code formatting tools (R styler).
-5. Bug fixes and code updates.
+4. Added pumping capacity scripts.
+5. Added code formatting tools (R styler).
+6. Bug fixes and code updates.
 
 ### Version 1.0.0 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3401411.svg)](https://doi.org/10.5281/zenodo.3401411)
 Includes all inputs for the anthropogenic impact modules as present in [VIC-WUR 2.0.0](https://github.com/wur-wsg/VIC/tree/VIC-WUR.2.0.0) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3399450.svg)](https://doi.org/10.5281/zenodo.3399450) (including scripts related to the VIC-WUR paper)